requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet>=3.0.0
pydantic-settings==2.2.1
python-dotenv==1.0.0
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
